Use `tini` with Kinesalite in order to handle SIGTERM
closes GH-5

This fixes a problem when using the Kinesalite image in for instance Docker
Compose, where the SIGTERM signal doesn't have any effect on a container
running the image. Docker Compose first sends a SIGTERM signal to containers
when running `docker-compose down` and then waits by default 10 seconds
before proceeding to send a SIGKILL signal. As a result of the SIGTERM
signal not being handled it always takes at least 10 seconds to stop the
Kinesalite container.

This commit fixes that by running the entrypoint script with `tini`, which
takes care of passing the signals to the child processes correctly.

Change-Id: Idea702c9e0e126eb7dc04f455a8bc454ce45f1db
Reviewed-on: https://gerrit.instructure.com/149656
Reviewed-by: Derek Bender <djbender@instructure.com>
Reviewed-by: Matthew Edwards <medwards@instructure.com>
Product-Review: Bryan Petty <bpetty@instructure.com>
QA-Review: Bryan Petty <bpetty@instructure.com>
Tested-by: Bryan Petty <bpetty@instructure.com>