Remove rust* images
The existing rust images are built against very out of date rust
versions and are not currently in use.  These are removed because they
are built on Trusty, and therefore approaching EOL.

The rust images can be restored and updated when they would be useful.

Refs TS-5292

Change-Id: Ie259940512981ae84644c83aae4f465cb7f0d8de
Reviewed-on: https://gerrit.instructure.com/179123
Tested-by: Jenkins
Reviewed-by: Tyler Pickett <tpickett@instructure.com>
Reviewed-by: Eric Coan <ecoan@instructure.com>
QA-Review: Ryan Taylor <rtaylor@instructure.com>