disable Passenger log prefix
By default, Passenger prefixes all logs with the PID and stream like:

    App PID stdout: ...

This breaks JSON-formatted logging, and is something we already always
strip back out using Fluentd in the logging pipeline since Passenger
wasn't capable of disabling this. As of Passenger 6.0.2, this
configuration option was added to disable it.

All docker images using Passenger are already upgraded to 6.0.6 (or
6.0.4 in the temporary regression image tag). Passenger defaults this to
off since that's how it's always behaved, but we just want to default it
to on here.

Change-Id: I8df2add29c53ca16d5b60b7d29cfd8ea73e6dbe5
Reviewed-on: https://gerrit.instructure.com/c/dockerfiles/+/250349
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Alex Slaughter <aslaughter@instructure.com>
Reviewed-by: Ryan Taylor <rtaylor@instructure.com>
QA-Review: Bryan Petty <bpetty@instructure.com>