Fix potential redirection/wrong answer issue within K8s envs
Reason: when doing healthchecks or even in-cluster cross-pod API requests, the traffic is expected to happen in HTTP. However, due to the way this rule:

if ($http_x_forwarded_proto != 'https') {
  return   301 https://$host$request_uri;
}

is written, it would cause a redirection even when the X-Forwarded-Proto header is not set up.

The fix corrects it by specifically looking for the 'http' value.

Change-Id: Ie18297fa024fbce6df2b1f62aa38eb9d5248862c
Reviewed-on: https://gerrit.instructure.com/c/dockerfiles/+/291555
Reviewed-by: James Belleau <james.belleau@instructure.com>
Reviewed-by: Jacob Burroughs <jburroughs@instructure.com>
QA-Review: Andras Kemeny <akemeny@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>